Create Symbol database (PDB) file for release DLL build
Symbol database (PDB) files for post-mortem debugging are not being added to the release DLL target because the /DEBUG switch is not added to the linker command line for release DLL builds. The /DEBUG switch is added for all build targets that specify GEN_PDB=yes. For DLL targets, the generated PDB file is moved from the \lib distribution folder to the \bin folder as it provides symbols for the DLL file, not the import library.