Make Curl_gss_log_error more verbose
Dispaly also the GSS_C_GSS_CODE (major code) when specified,
instead of only GSS_C_MECH_CODE (minor code).

In addition, the old code was printing twice a colon after the
prefix and also miscalculated the length of the buffer in between
calls to gss_display_status (the length of ": " was missing).

Also, gss_buffer is not guaranteed to be NULL terminated and
need to restrict reading by its length.