The error variable has to be reset before verifying each socket connection.
The error variable has to be reset before verifying each socket connection.
In other case, an error in the first socket might make Curl_is_connected
think the second socket in conn->tempsock has an error too and fail with
a "Network is unreachable" error in perfectly valid situations.

This fixes intermittent network errors from happening when ipv6 is not
available and the second choice is a valid ipv4 connection.