sendf: fix chop write for headers. draft3
use network encoded linefeed for parsing headers due to non ascii
character encoding.

sending headers the network encoded linefeed with the header. regardless
of whether or not we use a network encoded linefeed we have to do this.

on pause for buf param do not calculate buffer offset if buffer is null.