nss: work around race condition in PK11_FindSlotByName
Serialise the call to PK11_FindSlotByName in nss_create_object to avoid
spurious errors in a multi-threaded environment. The underlying cause
is a race condition in nssSlot_IsTokenPresent.

There are other paths into nssSlot_IsTokenPresent which may potentially
have the same problem, e.g. other calls to PK11_FindSlotByName, but this
is one path for which we have a test case.

Bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1297397