Retain original behaviour and improve file copying.
@captain-caveman2k
The easiest way to retain the original behaviour is to revert the change that removes the \DEBUG switch for debug DLL builds. If I make GEN_PDB default to true for debug DLL builds, then the compiler emits a warning for every object compilation because both \Z7 and \Zi are defined.

I've modified the file copying to avoid the rename.

I've also updated the comments to match the locations that the files end up in.