Avoid parsing certificates when not in verbose mode
The information extracted from the server certificates in step 3
is only used when in verbose mode, and there is no error handling
or validation performed as that has already been done. Only run
the certificate information extraction when in verbose mode and
avoid the extra processing with the CURL_DISABLE_VERBOSE_STRINGS
macro when not in verbose mode.