Winbuild (1201) Update build for OpenSSL
https://github.com/curl/curl/issues/1201

Introduce 5 new properties which can be set on the nmake command line

SSL_LIBS        - The lib names for the build.  Defaults to the pre 1.1 names.
                  Allows easy use of OpenSSL 1.1 names
SSL_HEADER_PATH - The path to the location of the OpenSSL header files.
                  Compatible with WITH_DEVEL
SSL_LIB_PATH    - The path to the location of the OpenSSL lib files.
                  Compatible with WITH_DEVEL
SSL_CFLAGS      - Allows extra settings for OpenSSL related compilation
SSL_LFLAGS      - Allows extra settings for OpenSSL related links

To Note:

- In theory SSL_xFlags makes the need SSL_yyy_PATH redundant.
- SSL_LFLAGS was already exposed, SSL_CFLAGS wasn't.
- "Documented" in the master make file.  Precise wording needs work.