Winbuild (1201) Update build parameterization for zlib
https://github.com/curl/curl/issues/1201

Introduce 5 new properties which can be set on the nmake command line

ZLIB_LIBS        - The lib names for the build.  Defaults to the pre 1.1 names.
                   Allows easy use of OpenSSL 1.1 names
ZLIB_HEADER_PATH - The path to the location of the OpenSSL header files.
                   Compatible with WITH_DEVEL
ZLIB_LIB_PATH    - The path to the location of the OpenSSL lib files.
                   Compatible with WITH_DEVEL
ZLIB_CFLAGS      - Allows extra settings for OpenSSL related compilation
ZLIB_LFLAGS      - Allows extra settings for OpenSSL related links

similar notes as for previous checkin