tool_operate: on second thought...
SQUASHME

- Change --proxy-{cacert,capath,crlfile} not to default to their regular
  non-proxy counterparts, since it's counterintuitive.

The CA path and certificate bundle may be set at compile-time, and if so
will be the default values in libcurl for those options. However by
defaulting for example --proxy-cacert if unset to --cacert if set that
would override the compile-time certificate file, which may not be what
was intended.

What is more intuitive is curl --cacert foo.pem and expecting that
since --proxy-cacert is unused then libcurl will use the default
certificate bundle set at compile time.