curl.h: look up curl headers in include path
When building the library, it would include the wrong configuration.
Test (where .. is srcdir) and curlbuild.h contains an error directive:

    In file included from ../include/curl/curl.h:38:0,
                     from /tmp/curl-dev/lib/cookie.h:26,
                     from /tmp/curl-dev/lib/urldata.h:82,
                     from /tmp/curl-dev/lib/url.c:70:
    ../include/curl/curlbuild.h:1:2: error: #error SHOULD NOT HAVE BEEN INCLUDED

To solve this issue, ensure that the header files looks in include
paths. This can potentially result in build failures when people use a
non-standard include path for curl and included curl.h directly
(contrary to the documented use of curl/curl.h).