macOS: connectx() function is weak linked to fix crash in 10.8
The connectx() function call appeared in Darwin 15.0.0
That covers OS X 10.11, iOS 9 and tvOS 9.

connectx() is not declared with weak_import attribute in sys/socket.h,
but it’s allowed to redeclare it with missing availability and weak_import attributes
which opens possibility to build libcurl that supports range of OS versions
with or without connectx() functionality.

Because libcurl uses -Wpartial-availability to warn about symbols not available
for all system targets, use of weak linked symbols need to be done within
pragma blocks isolating those specific cases.

[Bug: https://github.com/curl/curl/issues/1330]