include: curl/system.h is a run-time replacement for curlbuild.h
curl/system.h is used instead of curl/curlbuild.h

curl/system.h determines a few data sizes, data types and include file
status based on available preprocessor defines. This in order to avoid
relying on a build-time generated file that makes it complicated to do
32 and 64 bit bields from the same installed set of headers.

Define 'CURL_BUILDTIME' to instead use the old header if you need to,
and report this case back to the curl project.

curl/curlbuild.h will be removed completely at a later point in time
when we feel confident it isn't needed as a work-around any more.