SPNEGO - Use IP address if canonic. is enabled
If the target host we're connecting to is a round-robin DNS and the
authentication library will canonicalize the host (by doing a forward
and a reverse lookup), curl and gssapi will frequently disagree over
the Service Principal Name.

Introduce a new function in vauth.h to infer whether or not
canonicalization is happening in that particular mechanics.

For gssapi when using krb5, we implement the configuration profile
parsing to determine whether or not the "rdns" option in "libdefaults"
is set to false.

Every other implementation defaults to the current behavior of sending
the original hostname.

In theory, the correct thing would be to always give the ip address
unconditionally since according to the GSSAPI standard, the
canonicalization is not optional (and it is indeed a fundamental part
of the security model). However, that'd be a breaking change for
people that have set "rdns" to "false" in their krb5 configuration.

I am not familiar enough with the Windows SPI API to know whether or
not that is an option there.