multi: clarify condition in curl_multi_wait
`if(nfds || extra_nfds) {` is followed by `malloc(nfds * ...)`.

If `extra_fs` could be non-zero when `nfds` was zero, then we have `malloc(0)` which is allowed to return `NULL`. But, malloc returning NULL can be confusing. In this code, the next line would treat the NULL as an allocation failure.

It turns out, if `nfds` is zero then `extra_nfds` must also be zero.  The final value of `nfds` includes `extra_nfds`.  So the test for `extra_nfds` is redundant.  It can only confuse the reader.