test1521: test *all* curl_easy_setopt options
mk-lib1521.pl generates a test program (lib1521.c) that calls
curl_easy_setopt() for every known option with a few typical values to
make sure they work (ignoring the return codes).

Some small changes was necessary to avoid asserts and NULL accesses when
doing this.

The perl script is not automatically run but needs to be manually rerun
when we add new options.