non-ascii: allow conversion functions to be called with a NULL data handle.
This fixes segmentation faults, such as they can occur when calling
deprecated curl_escape() and curl_unescape() functions with a build that
enables character code conversions.
Using a NULL handle should be however avoided: a conversion descriptor is
created and deleted each time a conversion function is called with a NULL
handle, and thus is not very efficient.