openssl: improve data-pending check for https proxy
- Allow proxy_ssl to be checked for pending data even when connssl does
  not yet have an SSL handle.

This change is for posterity. Currently there doesn't seem to be a code
path that will cause a pending data check when proxyssl could have
pending data and the connssl handle doesn't yet exist *.

* Recall that an https proxy connection starts out in connssl but if the
destination is also https then the proxy SSL backend data is moved from
connssl to proxyssl, which means connssl handle is temporarily empty
until an SSL handle for the destination can be created.

Ref: https://github.com/curl/curl/commit/f4a6238#commitcomment-24396542

Closes #xxxx