hostip: allow debug builds to override what host name to resolve
The environment variable CURL_URL_HOSTNAME can be set to a host name
that will make libcurl resolve and use that host name instead of the
name it would otherwise use. Only for debug builds.

This also removes the special case debug support for LocalHost in the
interface string that kind of provided a similar debug approach. That
was brought in 7e07da977ce748 but doesn't seem to be used these days.