Content-Disposition: Add support for UTF-8 encoded filename* field.
If -J, --remote-header-name is specified, and filename* is given in Content-Disposition as utf-8 with a blank language tag, e.g. filename*=utf-8''foo%C2%AE, it is decoded by libcurl's curl_easy_unescape according to RFC 3986 and used as the filename for -O, --remote-name.  A UTF-8 filename* takes precedence over an ASCII filename (no asterisk) if both are specified.

Replacing memcmp with curl_strnequal also adds case insensitivity to filename and filename* parameters as is specified in RFC6266 section 4.3.

Closes #1888
Reported-by: devbazilio