curl_sspi: Use Unicode identity in non-UNICODE SSPI builds
- Make it so SSPI non-UNICODE builds can handle Unicode user/pass.

Prior to this change those builds basically attempted to create an
identity in the local codepage but that didn't work.

In more detail, what was expected to be a UTF-8 user/pass supplied by
the user was incorrectly treated as ANSI (ie user's local codepage),
which would cause AcquireCredentialsHandleA to fail at some later point
due to the incorrect encoding. Of note is even if the user tried ANSI
encoding for the user/pass (which may or may not be perceived as correct
since we don't document it), empirical testing has shown the conversion
would not necessarily be successful anyway (In other words it would be
incorrect to assume that internally Windows was doing a conversion from
ANSI -> UTF-16 like it does for most ANSI functions). See #2120 for more
information.

Ref: https://github.com/curl/curl/issues/2120