MacOSX-Framework: Added --without-ca-bundle to configure lines
Adding --without-ca-bundle to the configure lines in the MacOSX-Framework
script brings the functionality of what to do with a self-signed certificate
that has been trusted in a keychain (e.g. login or System) into line with the
curl installation that comes with macOS does which is not to look at
/etc/ssl/cert.pem based on the curl_config.h generated in curl-95.70.1 (10.12)
on the Apple Open Source website.

Before adding --without-ca-bundle, running ./MacOSX-Framework and testing
with 'curl https://<trusted self signed domain>' in the src directory
returned error message 'curl: (51) SSL: certificate verification failed
(result: 5)'. Error code 5 is kSecTrustResultRecoverableTrustFailure
which shouldn't be happening if it is looking at the keychain(s) properly and
a trusted self-signed certificate is there.

Tested on macOS Sierra using the included installation (7.54.0) and latest
development build (7.58.0-DEV) at commit 9c6a6be88.